backend/requirements.txt
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
async-timeout==5.0.1
boto3==1.42.59
botocore==1.42.59
certifi==2026.2.25
charset-normalizer==3.4.4
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
jmespath==1.1.0
python-jobspy
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.83
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
lxml==6.0.2
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0




